+++ tests/fixtures/config.py
@@ -1,3 +1,12 @@
 # -*- coding: utf-8 -*-
 
 EmptyConfig = {}
+
+SomeConfig = {
+    "name": "Jane",
+    "age": 34,
+    "path": {
+        "type": "env",
+        "key": "PATH"
+    }
+}